this is to address #16 . it goes with corresponding changes to https://github.com/brentp/vcfgo and https://github.com/brentp/bix
when using 'self' with a vcf annotation file, it will set the
Number in the annotated query file to what it was in the annotation
file.

There are some additional problems that are fairly rare, but common
enough that we may need a solution. I will add more info to #16.