Make header case-insensitive
If you go to the CADD website (http://cadd.gs.washington.edu/) you can submit variants to be scored (useful for novel indels not in their pre-scored database). The score file you get back is in the same TSV format as their database except for the casing of the header line (CHROM instead of Chrom, POS instead of pos, etc.). This change makes it so the script can parse either type of score file.
I'm not sure if you'd want to make this change or not or if there's a better way to do it, but I thought I'd offer it up as a suggestion.