better warnings and handling of multiple alts
re #83 and #87

if there are already values in the query info field for a variant with
multiple alternates, incoming values will only overwrite existing values
if they are non-nil (or non-zero values of the type).

thanks @RoanKanninga for reporting and providing test-cases.

when Number=1 in the annotation file (and therefore the input file)
and there are multiple alternates in the input file, the values
can be out of order. This now issues a warning indicating the file
and the field in question and noting that it can be mitigated by
decomposing the input file.